rename the tab3.tsx into calculator.tsx and also its code
