+++ src/App.tsx
@@ -17,7 +17,7 @@ import Home from './pages/home';
 
 import Profile from './pages/Profile';
 import Clickcounter from './pages/Clickcounter';
-import Tab3 from './pages/Tab3';
+import Tab3 from './pages/Calculator';
 
 /* Core CSS required for Ionic components to work properly */
 import '@ionic/react/css/core.css';
